+++ CHANGELOG.md
@@ -8,6 +8,7 @@ FEATURES:
 * Set device network info
 * Set device SN (only on supported Anviz devices)
 * Set device ID
+* Reboot device
 
 BUG FIXES: